+++ yggdrasil_decision_forests/port/python/ydf/model/generic_model.py
@@ -482,6 +482,60 @@ Use `model.describe()` for more details
     """
     raise NotImplementedError
 
+  def predict_shap(
+      self,
+      data: dataset.InputDataset,
+      *,
+      num_threads: Optional[int] = None,
+  ) -> Tuple[Dict[str, np.ndarray], np.ndarray]:
+    """Returns the SHAP value of the model for each example in the dataset.
+
+    Usage example:
+
+    ```python
+    import pandas as pd
+    import ydf
+
+    # Train model
+    train_ds = pd.read_csv("train.csv")
+    model = ydf.RandomForestLearner(label="label").train(train_ds)
+
+    # Computes the SHAP values on the test dataset.
+    test_ds = pd.read_csv("test.csv")
+    shap_values, initial_value = model.predict_shap(test_ds)
+    ```
+
+    The shap values (`shap_values`) is a dictionary mapping feature names to a
+    float32 Numpy array of shape [num examples, num output] or [num examples]
+    (if the model has a single output).
+
+    The second returned value `initial_value` is a float32 Numpy array of shape
+    [num output] or [] (if the model has a single output) with the initial
+    (a.k.a. offset predictions).
+
+    The prediction of the model (computed with `model.predict`) is equal to
+    the `initial_value` plus all the shape values `shap_values`. Note that for
+    models with an activation function (a.k.a. linkage funciton; e.g. sigmoid on
+    binary classification GBT models), the SHAP values are computed before the
+    activation function (e.g., on the logits on binary classification GBT
+    models).
+
+    The SHAP `initial_value` are generally not equal to the "initial prediction"
+    of some models (e.g. gradient boosted trees).
+
+    Args:
+      data: Dataset. Supported formats: VerticalDataset, (typed) path, list of
+        (typed) paths, Pandas DataFrame, Xarray Dataset, TensorFlow Dataset,
+        PyGrain DataLoader and Dataset (experimental, Linux only), dictionary of
+        string to NumPy array or lists. If the dataset contains the label
+        column, that column is ignored.
+      num_threads: Number of threads used to run the model.
+
+    Returns:
+      Dictionary of shape values and initial model values.
+    """
+    raise NotImplementedError("SHAP is not implemented for this model")
+
   @abc.abstractmethod
   def evaluate(
       self,
@@ -1522,6 +1576,27 @@ class GenericCCModel(GenericModel):
           ds._dataset, use_slow_engine, num_threads=num_threads  # pylint: disable=protected-access
       )
     return result
+
+  def predict_shap(
+      self,
+      data: dataset.InputDataset,
+      *,
+      num_threads: Optional[int] = None,
+  ) -> Tuple[Dict[str, np.ndarray], np.ndarray]:
+    if num_threads is None:
+      num_threads = concurrency.determine_optimal_num_threads(training=False)
+
+    with log.cc_log_context():
+      # The data spec contains the label / weights /  ranking group / uplift
+      # treatment column, but those are not required for making predictions.
+      ds = dataset.create_vertical_dataset(
+          data,
+          data_spec=self._model.data_spec(),
+          required_columns=self.input_feature_names(),
+      )
+      return self._model.PredictShap(
+          ds._dataset, num_threads=num_threads  # pylint: disable=protected-access
+      )
 
   def evaluate(
       self,