+++ yggdrasil_decision_forests/port/python/ydf/model/model.cc
@@ -123,6 +123,9 @@ void init_model(py::module_& m) {
            py::arg("dataset"), py::arg("use_slow_engine"),
            py::arg("num_threads"))
       // WARNING: This method releases the Global Interpreter Lock.
+      .def("PredictShap", WithStatusOr(&GenericCCModel::PredictShap),
+           py::arg("dataset"), py::arg("num_threads"))
+      // WARNING: This method releases the Global Interpreter Lock.
       .def("Evaluate", WithStatusOr(&GenericCCModel::Evaluate),
            py::arg("dataset"), py::arg("options"), py::arg("weighted"),
            py::arg("label_col_idx"), py::arg("group_col_idx"),